+++ src/UI/SignTransactionsModals/SignWithLedgerModal/SignWithLedgerModal.tsx
@@ -1,8 +1,23 @@
 import React from 'react';
+import { useGetAccountInfo } from 'hooks';
 import { SignModalPropsType } from 'types';
+import { getLedgerVersionOptions } from 'utils';
 import { SignWithDeviceModal } from '../SignWithDeviceModal';
+import { LedgerGuardianSigningModalError } from './LedgerGuardianSigningModalError';
 
 export const SignWithLedgerModal = (props: SignModalPropsType) => {
+  const {
+    ledgerAccount,
+    account: { isGuarded }
+  } = useGetAccountInfo();
+  const { ledgerWithGuardians } = getLedgerVersionOptions(
+    ledgerAccount?.version ?? ''
+  );
+
+  if (isGuarded && !ledgerWithGuardians) {
+    return <LedgerGuardianSigningModalError {...props} />;
+  }
+
   return (
     <SignWithDeviceModal
       {...props}