+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Prevent ledger signing of guarded transactions when MultiversX App is too old](https://github.com/multiversx/mx-sdk-dapp/pull/746)
 - [Add batch transactions tracking mechanism](https://github.com/multiversx/mx-sdk-dapp/pull/745)
 
 ## [[v2.13.0]](https://github.com/multiversx/mx-sdk-dapp/pull/770)] - 2023-05-09